requirements.txt
# NBABot v10.0 Dependencies

# Discord
discord.py>=2.3.0

# HTTP Client
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# Development (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
